examples/reference/requirements.txt
scikit-learn==0.23.1
python-mnist
dvc[gs]
google-cloud-aiplatform
google-cloud-storage
pymongo
sacred
starlette
uvicorn
dill==0.3.4
scipy==1.6.3
google-cloud-logging
google-cloud-secret-manager